// src/app/enums/AppRouting.enum.ts
export const enum  AppRouting {
    Login = 'Login',
    Home = 'Home',
    Profile = 'Profile',
    UserList= 'Profile/all',
    Office = 'Office',
    OfficeList = 'Office/all',
    Stock = 'Stock',
    Dispatch = 'Dispatch',
    DispatchList = 'Dispatch/all',
    DispatchCreate = 'Dispatch/create',
    DispatchRecive = 'Dispatch/recive',
    History = 'History',
    Sale = 'Sale',
    Discount ='Discount',
    DiscountCreate = 'Discount/create',
    Changes = 'Changes'

}